requirements.txt
altair @ file:///C:/b/abs_0b1avjyzoc/croot/altair_1743016738596/work
arch==8.0.0
asttokens==3.0.0
attrs @ file:///C:/miniconda3/conda-bld/attrs_1762356899359/work
beautifulsoup4 @ file:///C:/b/abs_95vq944bnw/croot/beautifulsoup4-split_1756453107890/work
blinker @ file:///C:/b/abs_b1i87khtob/croot/blinker_1737448732095/work
Bottleneck @ file:///C:/b/abs_816hr2khp1/croot/bottleneck_1731058648110/work
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
brotlicffi @ file:///C:/miniconda3/conda-bld/brotlicffi_1761745219545/work
cachetools @ file:///C:/b/abs_fcvwpc5vh9/croot/cachetools_1738224687938/work
certifi @ file:///C:/miniconda3/conda-bld/certifi_1759786076808/work/certifi
cffi @ file:///C:/miniconda3/conda-bld/cffi_1761832792955/work
charset-normalizer==3.4.3
click @ file:///C:/b/abs_bbg7qqz4dk/croot/click_1754477951021/work
cmdstanpy==1.3.0
colorama==0.4.6
comm==0.2.2
contourpy @ file:///C:/b/abs_f2u2o_8s9g/croot/contourpy_1732540071787/work
curl_cffi==0.13.0
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Cython==3.2.1
debugpy==1.8.13
decorator==5.2.1
et_xmlfile @ file:///C:/b/abs_b0ifp_j3js/croot/et_xmlfile_1756462499666/work
executing==2.2.0
filelock==3.20.0
fonttools @ file:///C:/miniconda3/conda-bld/fonttools_1758791708108/work
frozendict==2.4.6
gitdb @ file:///C:/b/abs_1bkmriimss/croot/gitdb_1753362133190/work
GitPython @ file:///C:/miniconda3/conda-bld/gitpython_1761291338863/work
google @ file:///home/conda/feedstock_root/build_artifacts/googlesearch_1735279112531/work
greenlet==3.2.4
holidays==0.85
html5lib @ file:///Users/ktietz/demo/mc3/conda-bld/html5lib_1629144453894/work
idna==3.10
importlib_resources==6.5.2
investpy==1.0.8
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
joblib @ file:///C:/miniconda3/conda-bld/joblib_1757926362684/work
jsonschema @ file:///C:/miniconda3/conda-bld/jsonschema_1762459308681/work
jsonschema-specifications @ file:///C:/miniconda3/conda-bld/jsonschema-specifications_1762788613133/work
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver @ file:///C:/b/abs_faf90xet7a/croot/kiwisolver_1737040915779/work
lightgbm==4.6.0
llvmlite @ file:///C:/miniconda3/conda-bld/llvmlite_1760469323312/work
lxml==5.3.0
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
matplotlib==3.10.6
matplotlib-inline==0.1.7
mkl-service==2.5.2
mkl_fft @ file:///C:/b/abs_64riv3lfyn/croot/mkl_fft_1751310924052/work
mkl_random @ file:///C:/b/abs_47jvsmau4j/croot/mkl_random_1751310922820/work
multitasking==0.0.12
mysql-connector-python==9.5.0
narwhals @ file:///C:/miniconda3/conda-bld/narwhals_1760003513313/work
nest-asyncio==1.6.0
numba @ file:///C:/miniconda3/conda-bld/numba_1760991948802/work
numexpr @ file:///C:/b/abs_3buvfgcf6r/croot/numexpr_1755766488094/work
numpy @ file:///C:/miniconda3/conda-bld/numpy_and_numpy_base_1757670256083/work/dist/numpy-2.3.3-cp313-cp313-win_amd64.whl#sha256=d9474e66189cba52ea8e11316ddea0a38ae7f8fc06e3ef798d9bbfddc6ba51a5
openpyxl @ file:///C:/b/abs_6cdczthntz/croot/openpyxl_1736371239628/work
packaging==24.2
pandas @ file:///C:/b/abs_58a1hlvjeo/croot/pandas_1756466410850/work/dist/pandas-2.3.2-cp313-cp313-win_amd64.whl#sha256=e3d25986bd6f7ec53acbba8bf94ee4f49c11b115835d0b9a4c62809e4f788a80
pandas-ta @ file:///home/conda/feedstock_root/build_artifacts/pandas-ta_1758611198455/work
parso==0.8.4
patsy==1.0.2
peewee==3.18.2
pillow @ file:///C:/b/abs_b06nvnir5m/croot/pillow_1752524543576/work
platformdirs==4.3.6
plotly @ file:///home/conda/feedstock_root/build_artifacts/plotly_1762283526932/work
pmdarima==2.1.1
prompt_toolkit==3.0.50
prophet==1.2.1
protobuf @ file:///C:/b/abs_b5v84rwh56/croot/protobuf_1742419721750/work/bazel-bin/python/dist/protobuf-5.29.3-cp313-abi3-win_amd64.whl#sha256=0ef6aadfdcd9f988cbfc2d48eb8f7ed3d64acc158314c82a3a9c456f3fcc53a4
psutil==7.0.0
pure_eval==0.2.3
pyarrow @ file:///C:/miniconda3/conda-bld/pyarrow_1759833600190/work/python
pycparser @ file:///C:/miniconda3/conda-bld/pycparser_1757496153123/work
pydeck @ file:///C:/b/abs_47xw5z_59x/croot/pydeck_1738340983078/work
Pygments==2.19.1
pyparsing @ file:///C:/Users/dev-admin/buildout/croot/pyparsing_1735861809022/work
PyQt6==6.7.1
PyQt6_sip @ file:///C:/b/abs_25j3xlesmj/croot/pyqt-split_1753427295085/work/pyqt_sip
PySocks @ file:///C:/miniconda3/conda-bld/pysocks_1761753030965/work
python-dateutil @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/python-dateutil_1729038406576/work
python-dotenv==1.2.1
pytz @ file:///C:/b/abs_f8wdzeix0n/croot/pytz_1752135878094/work
pywin32==310
PyYAML @ file:///C:/b/abs_79lqlbkifj/croot/pyyaml_1731006239057/work
pyzmq==26.3.0
referencing @ file:///C:/miniconda3/conda-bld/referencing_1762536906371/work
requests @ file:///C:/miniconda3/conda-bld/requests_1762359611326/work
rpds-py @ file:///C:/miniconda3/conda-bld/rpds-py_1762366664211/work
scikit-learn @ file:///C:/miniconda3/conda-bld/scikit-learn_1758619682247/work
scipy @ file:///C:/miniconda3/conda-bld/scipy_1757315814389/work/dist/scipy-1.16.1-cp313-cp313-win_amd64.whl#sha256=5d210da8b34d75d142b4253e4453cc8e62ddf11e3038b472f8e308d0c332f768
setuptools==72.1.0
sip @ file:///C:/b/abs_5cto136kse/croot/sip_1738856220313/work
six @ file:///C:/b/abs_149wuyuo1o/croot/six_1744271521515/work
smmap @ file:///tmp/build/80754af9/smmap_1611694433573/work
soupsieve @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/soupsieve_1729039543636/work
SQLAlchemy==2.0.44
stack-data==0.6.3
stanio==0.5.1
statsmodels==0.14.5
streamlit @ file:///home/conda/feedstock_root/build_artifacts/streamlit_1761776807357/work
tabulate @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/tabulate_1729056893645/work
tenacity @ file:///C:/miniconda3/conda-bld/tenacity_1760375744772/work
threadpoolctl @ file:///C:/b/abs_3904pspv77/croot/threadpoolctl_1731006105881/work
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tornado==6.4.2
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
traitlets==5.14.3
typing_extensions @ file:///C:/b/abs_ecq8gc0vbm/croot/typing_extensions_1756281142218/work
tzdata @ file:///croot/python-tzdata_1746123641790/work
Unidecode==1.4.0
urllib3 @ file:///C:/b/abs_393g1ayee_/croot/urllib3_1750775481226/work
watchdog @ file:///C:/miniconda3/conda-bld/watchdog_1759775798728/work
wcwidth==0.2.13
webencodings @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/webencodings_1729039571130/work
websockets==15.0.1
wheel==0.45.1
win_inet_pton @ file:///C:/miniconda3/conda-bld/win_inet_pton_1761746278300/work
xgboost==3.1.2
yfinance==0.2.66
